Split functions to create a package.
